Refactor parameter flattening to generalized util
